+++ test/riemann_discovery/file_test.clj
@@ -63,31 +63,31 @@
       (let [events (first (last @calls))]
         (is (= (count events) 5))
         (is (some #{{:host "kafka1"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "kafka2"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "api1"
-                     :service "api"
+                     :service "api-discovery"
                      :ttl 120
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "zookeeper1"
-                     :service "zookeeper"
+                     :service "zookeeper-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "cassandra1"
-                     :service "cassandra"
+                     :service "cassandra-discovery"
                      :ttl 120
                      :time 10
                      :state "added"
@@ -110,19 +110,19 @@
       (let [events (first (last @calls))]
         (is (= (count events) 3))
         (is (some #{{:host "kafka1"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "kafka2"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "zookeeper1"
-                     :service "zookeeper"
+                     :service "zookeeper-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
@@ -135,13 +135,13 @@
       (let [events (first (last @calls))]
         (is (= (count events) 2))
         (is (some #{{:host "cassandra1"
-                     :service "cassandra"
+                     :service "cassandra-discovery"
                      :ttl 120
                      :time 310
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "api1"
-                     :service "api"
+                     :service "api-discovery"
                      :ttl 120
                      :time 310
                      :state "added"