+++ test/riemann_discovery/http_test.clj
@@ -57,25 +57,25 @@
       (let [events (first (last @calls))]
         (is (= (count events) 4))
         (is (some #{{:host "kafka1"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "kafka2"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "api1"
-                     :service "api"
+                     :service "api-discovery"
                      :ttl 120
                      :time 10
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "zookeeper1"
-                     :service "zookeeper"
+                     :service "zookeeper-discovery"
                      :ttl 60
                      :time 10
                      :state "added"
@@ -98,19 +98,19 @@
       (let [events (first (last @calls))]
         (is (= (count events) 3))
         (is (some #{{:host "kafka1"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "kafka2"
-                     :service "kafka"
+                     :service "kafka-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
                      :tags ["riemann-discovery"]}} events))
         (is (some #{{:host "zookeeper1"
-                     :service "zookeeper"
+                     :service "zookeeper-discovery"
                      :ttl 60
                      :time 190
                      :state "added"
@@ -123,7 +123,7 @@
       (let [events (first (last @calls))]
         (is (= (count events) 1))
         (is (some #{{:host "api1"
-                     :service "api"
+                     :service "api-discovery"
                      :ttl 120
                      :time 310
                      :state "added"